Move audio module to its own crate
