+++ examples/Cargo.toml
@@ -5,7 +5,7 @@ license = "MPL-2.0"
 
 [dependencies]
 rand = "0.5.0"
-servo_media = { path = "../servo-media" }
+servo-media = { path = "../servo-media" }
 
 [[bin]]
 name = "audio_decoder"
