+++ servo-media-audio/src/node.rs.orig
@@ -1,0 +1,182 @@
+<<<<<<< HEAD
+use audio::param::{Param, ParamType};
+||||||| parent of 8a3d52e... Switch over to ParamIds for everything
+use audio::param::Param;
+use audio::param::ParamId;
+=======
+use audio::param::{Param, ParamType, UserAutomationEvent};
+>>>>>>> 8a3d52e... Switch over to ParamIds for everything
+use audio::channel_node::ChannelNodeOptions;
+use audio::block::{Chunk, Tick};
+use audio::buffer_source_node::{AudioBufferSourceNodeMessage, AudioBufferSourceNodeOptions};
+use audio::gain_node::GainNodeOptions;
+use audio::oscillator_node::OscillatorNodeOptions;
+
+/// Type of AudioNodeEngine.
+#[derive(Debug, Clone)]
+pub enum AudioNodeType {
+    AnalyserNode,
+    BiquadFilterNode,
+    AudioBuffer,
+    AudioBufferSourceNode(AudioBufferSourceNodeOptions),
+    ChannelMergerNode(ChannelNodeOptions),
+    ChannelSplitterNode(ChannelNodeOptions),
+    ConstantSourceNode,
+    ConvolverNode,
+    DelayNode,
+    DestinationNode,
+    DynamicsCompressionNode,
+    GainNode(GainNodeOptions),
+    IIRFilterNode,
+    OscillatorNode(OscillatorNodeOptions),
+    PannerNode,
+    PeriodicWave,
+    ScriptProcessorNode,
+    StereoPannerNode,
+    WaveShaperNode,
+}
+
+#[derive(Copy, Clone, PartialEq, Eq, Debug)]
+pub enum ChannelCountMode {
+    Max,
+    ClampedMax,
+    Explicit
+}
+
+
+#[derive(Copy, Clone, PartialEq, Eq, Debug)]
+pub enum ChannelInterpretation {
+    Discrete,
+    Speakers
+}
+
+#[derive(Copy, Clone)]
+pub struct BlockInfo {
+    pub sample_rate: f32,
+    pub frame: Tick,
+    pub time: f64,
+}
+
+impl BlockInfo {
+    /// Given the current block, calculate the absolute zero-relative
+    /// tick of the given tick
+    pub fn absolute_tick(&self, tick: Tick) -> Tick {
+        self.frame + tick
+    }
+}
+
+
+pub struct ChannelInfo {
+    pub count: u8,
+    pub mode: ChannelCountMode,
+    pub interpretation: ChannelInterpretation,
+}
+
+impl Default for ChannelInfo {
+    fn default() -> Self {
+        ChannelInfo {
+            count: 2,
+            mode: ChannelCountMode::Max,
+            interpretation: ChannelInterpretation::Speakers,
+        }
+    }
+}
+
+
+pub trait AudioNodeCommon {
+    fn channel_info(&self) -> &ChannelInfo;
+
+    fn channel_info_mut(&mut self) -> &mut ChannelInfo;
+}
+
+/// This trait represents the common features of all audio nodes.
+pub trait AudioNodeEngine: Send + AudioNodeCommon {
+    fn node_type(&self) -> &'static str;
+
+    fn process(&mut self, inputs: Chunk, info: &BlockInfo) -> Chunk;
+
+    fn message(&mut self, msg: AudioNodeMessage, sample_rate: f32) {
+        match msg {
+            AudioNodeMessage::SetChannelCount(c) => self.set_channel_count(c),
+            AudioNodeMessage::SetChannelMode(c) => self.set_channel_count_mode(c),
+            AudioNodeMessage::SetChannelInterpretation(c) => self.set_channel_interpretation(c),
+            AudioNodeMessage::SetParam(id, event) => {
+                self.get_param(id).insert_event(event.to_event(sample_rate))
+            }
+            _ => self.message_specific(msg, sample_rate),
+        }
+    }
+
+    /// Messages specific to this node
+    fn message_specific(&mut self, _: AudioNodeMessage, _sample_rate: f32) {}
+
+    fn input_count(&self) -> u32 {
+        1
+    }
+    fn output_count(&self) -> u32 {
+        1
+    }
+
+    /// Number of input channels for each input port
+    fn channel_count(&self) -> u8 {
+        self.channel_info().count
+    }
+
+    fn channel_count_mode(&self) -> ChannelCountMode {
+        self.channel_info().mode
+    }
+
+    fn channel_interpretation(&self) -> ChannelInterpretation {
+        self.channel_info().interpretation
+    }
+
+    fn set_channel_interpretation(&mut self, i: ChannelInterpretation) {
+        self.channel_info_mut().interpretation = i
+    }
+    fn set_channel_count(&mut self, c: u8) {
+        self.channel_info_mut().count = c;
+    }
+    fn set_channel_count_mode(&mut self, m: ChannelCountMode) {
+        self.channel_info_mut().mode = m;
+    }
+
+    /// If we're the destination node, extract the contained data
+    fn destination_data(&mut self) -> Option<Chunk> {
+        None
+    }
+
+    fn get_param(&mut self, _: ParamType) -> &mut Param {
+        panic!("No params on node {}", self.node_type())
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum AudioNodeMessage {
+    AudioBufferSourceNode(AudioBufferSourceNodeMessage),
+    AudioScheduledSourceNode(AudioScheduledSourceNodeMessage),
+    SetChannelCount(u8),
+    SetChannelMode(ChannelCountMode),
+    SetChannelInterpretation(ChannelInterpretation),
+    SetParam(ParamType, UserAutomationEvent)
+}
+
+/// This trait represents the common features of the source nodes such as
+/// AudioBufferSourceNode, ConstantSourceNode and OscillatorNode.
+/// https://webaudio.github.io/web-audio-api/#AudioScheduledSourceNode
+pub trait AudioScheduledSourceNode {
+    /// Schedules a sound to playback at an exact time.
+    /// Returns true if the scheduling request is processed succesfully.
+    fn start(&mut self, tick: Tick) -> bool;
+    /// Schedules a sound to stop playback at an exact time.
+    /// Returns true if the scheduling request is processed successfully.
+    fn stop(&mut self, tick: Tick) -> bool;
+}
+
+/// Type of message directed to AudioScheduledSourceNodes.
+#[derive(Debug, Clone)]
+pub enum AudioScheduledSourceNodeMessage {
+    /// Schedules a sound to playback at an exact time.
+    Start(f64),
+    /// Schedules a sound to stop playback at an exact time.
+    Stop(f64),
+}
