+++ servo-media-gstreamer/Cargo.toml
@@ -20,5 +20,5 @@ version = "0.11"
 [dependencies.num-traits]
 version = "0.1"
 
-[dependencies.servo_media]
-path = "../servo-media"
+[dependencies.servo-media-audio]
+path = "../servo-media-audio"
