+++ servo-media-gstreamer/src/lib.rs
@@ -1,13 +1,15 @@
-use servo_media::audio::context::AudioBackend;
+
 
 extern crate gstreamer_app as gst_app;
 extern crate gstreamer_audio as gst_audio;
 extern crate gstreamer as gst;
 
-extern crate servo_media;
+extern crate servo_media_audio;
 
 extern crate num_traits;
 extern crate byte_slice_cast;
+
+use servo_media_audio::context::AudioBackend;
 
 pub mod audio_decoder;
 pub mod audio_sink;
@@ -22,5 +24,8 @@ impl AudioBackend for GStreamerBackend {
     }
     fn make_sink() -> Result<Self::Sink, ()> {
         audio_sink::GStreamerAudioSink::new()
+    }
+    fn init() {
+        gst::init().unwrap();
     }
 }
