+++ servo-media/Cargo.toml
@@ -1,27 +1,14 @@
 [package]
-authors = ["Fernando Jiménez Moreno <ferjmoreno@gmail.com>"]
-license = "MPL-2.0"
-name = "servo_media"
+name = "servo-media"
 version = "0.1.0"
-
-[build-dependencies]
-regex = "1.0"
-zip = "0.3.1"
-
-[dependencies]
-smallvec = "0.6.1"
-servo_media_derive = { path = "../servo-media-derive" }
-
-[dependencies.petgraph]
-version = "0.4.12"
-features = ["stable_graph"]
-
-[dependencies.byte-slice-cast]
-version = "0.1"
-
-
-[dependencies.num-traits]
-version = "0.1"
+authors = ["Manish Goregaokar <manishsmail@gmail.com>"]
 
 [lib]
 name = "servo_media"
+
+[dependencies.servo-media-audio]
+path = "../servo-media-audio"
+
+
+[dependencies.servo-media-gstreamer]
+path = "../servo-media-gstreamer"
